Refactor links to urls and copyright into components folder and removed unused footer
