+++ src/App.js
@@ -4,7 +4,8 @@ import MainPage from './logged_in/components/layout/MainPage';
 import Login from "./logged_out/Login";
 import SignUp from "./logged_out/SignUp";
 import PrivateRoute from "./shared/components/PrivateRoute";
-import {LOG_IN_LINK, ROOT_LINK, SIGN_UP_LINK} from "./shared/costants/links";
+import {EDITOR_URL, LOG_IN_URL, MY_GRABITS_URL, ROOT_URL, SIGN_UP_URL} from "./shared/costants/urls";
+import GrabitPage from "./logged_in/components/layout/GrabitsPage";
 
 function NotFound() {
     return<h1>404 Not Found</h1>; // TODO: temporary component, do a proper 404 page
@@ -14,9 +15,11 @@ function App() {
     return (
         <Router>
             <Switch>
-                <PrivateRoute exact path={ROOT_LINK} component={MainPage}/>
-                <Route path={LOG_IN_LINK} component={Login}/>
-                <Route path={SIGN_UP_LINK} component={SignUp}/>
+                <PrivateRoute exact path={ROOT_URL} component={MainPage}/>
+                <PrivateRoute path={MY_GRABITS_URL} component={GrabitPage}/>
+                <PrivateRoute path={EDITOR_URL} component={MainPage}/>
+                <Route path={LOG_IN_URL} component={Login}/>
+                <Route path={SIGN_UP_URL} component={SignUp}/>
                 <Route component={NotFound}/>
             </Switch>
         </Router>
