--- src/logged_in/components/layout/Footer.js
@@ -1,43 +1,0 @@
-import React from 'react';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
-import Container from '@material-ui/core/Container';
-import Link from '@material-ui/core/Link';
-
-function Copyright() {
-  return (
-    <Typography variant="body2" color="textSecondary">
-      {'Copyright © '}
-      <Link color="inherit" href="https://gisilo.com/">
-        Gisilo
-      </Link>{' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  );
-}
-
-const useStyles = makeStyles((theme) => ({
-  footer: {
-    padding: theme.spacing(3, 2),
-    marginTop: 'auto',
-    backgroundColor:
-      theme.palette.type === 'light' ? theme.palette.grey[200] : theme.palette.grey[800],
-  },
-}));
-
-export default function StickyFooter() {
-  const classes = useStyles();
-
-  return (
-    <div>
-      <footer className={classes.footer}>
-        <Container maxWidth="sm">
-          <Typography variant="body1">Grask - a Gisilo software.</Typography>
-          <Copyright />
-        </Container>
-      </footer>
-    </div>
-  );
-}
