--- src/shared/costants/links.js
@@ -1,3 +1,0 @@
-export const ROOT_LINK = "/";
-export const SIGN_UP_LINK = "/signup";
-export const LOG_IN_LINK = "/login";
