+++ src/shared/costants/urls.js
@@ -1,0 +1,5 @@
+export const ROOT_URL = "/";
+export const MY_GRABITS_URL = "/grabits";
+export const EDITOR_URL = "/editor";
+export const SIGN_UP_URL = "/signup";
+export const LOG_IN_URL = "/login";
